Cookie\Jar: move tests to dedicated file
The tests for the `Cookie\Jar` class were up to now mixed in with the tests for the `Cookie` class.

This moves the `Jar` tests to a dedicated test file.

The actual test code has not been touched.

This includes duplicating a helper method. At a later point in time when these tests are reviewed, it can be decided whether this helper should move to the `TestCase`.
